% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrangeStatistics}
\alias{arrangeStatistics}
\title{arrangeStatistics}
\usage{
arrangeStatistics(env, effectsMatrix, names, intercept, M, D, S)
}
\arguments{
\item{env}{global environment where the environment of statistics is ('stats')}

\item{effectsMatrix}{is a matrix of effects from the processed formula object}

\item{names}{is the vector of row names from the matrix of effects (effectsMatrix)}

\item{intercept}{is a TRUE/FALSE value whether the intercept is specified or not}

\item{M}{is the number of events}

\item{D}{is the number of dyads (without event type)}

\item{S}{is the final number of predictors in the model}
}
\value{
three dimensional array of statistics and vector of names (slices)
}
\description{
arrangeStatistics
}
