% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{bma}
\usage{
bma(
  smm = NULL,
  weights = c("BIC", "WAIC", "pseudoBMAplus", "stacking", "all"),
  n_knots = 300,
  nsim = 5000
)
}
\arguments{
\item{smm}{can be either an object of class 'smm' or a list of objects of class 'smm'. The second option allows to combine multiple smm objects provided that they refer to the same model specification but to different sets of stepwise memory models.}

\item{weights}{vector of weighting systems to be used in the averaging. "BIC" performs the method were smm's are considered to be generative models for the posterior distribution (which might be an unrealistic assumption since the posterior trend can be smooth); The "pseudoBMAplus" and "stacking" are based on the measure of ELPD. "WAIC" is an approximation of the ELPD and will be . If "all", then all the available method will be used. Also a few methods can be specified}

\item{n_knots}{xx}

\item{nsim}{xx}
}
\value{
object of class 'bma'
}
\description{
A function which return an object of class 'bma' with the estimate of the posterior distribution of effects according to specific weighting systems
}
