% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremory-package.R
\docType{package}
\name{bremory-package}
\alias{bremory-package}
\alias{_PACKAGE}
\title{bremory: Modeling memory decay in relational event networks}
\description{
This package offers several approaches (semi-parametric and parametric) to model memory decay in relational event networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TilburgNetworkGroup/bremory}
  \item Report bugs at \url{https://github.com/TilburgNetworkGroup/bremory/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Arena \email{g.arena@tilburguniversity.edu}

Other contributors:
\itemize{
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
