% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{bremory}
\alias{bremory}
\title{bremory: A package for learning the memory retention process of actors in relational event histories.}
\description{
The bremory package provides a set of functions to estimate the posterior trend of effects
following the theoretical considerations in the relative paper 'title of the paper'
}
\section{\code{bremory} functions}{

The bremory functions ...
}

