% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEdgelist.R
\name{convertEdgelist}
\alias{convertEdgelist}
\title{convertEdgelist (edgelist) to be coded in Rcpp only for rem = TRUE and remember that the output indices must go from 1 to D (0 must not be as idyad index)}
\usage{
convertEdgelist(edgelist, riskset, actors, rem = FALSE)
}
\arguments{
\item{edgelist}{matrix of dimensions M*3  (by row[])}

\item{riskset}{matrix of dimensions N*(N-1) * (sender,receiver)}

\item{actors}{dataframe of actors names and corresponding integer id's.}

\item{rem}{boolean to understand whether the format of the edgelist has to be the 'rem' one (dyad,time) or the 'default' one (time,sender,receiver)}
}
\value{
converted edgelist according to 'rem' argument
}
\description{
A function which return the dimension of a specific environment. [remove export]
}
