% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertToReleventEdgelist}
\alias{convertToReleventEdgelist}
\title{convertToReleventEdgelist()}
\usage{
convertToReleventEdgelist(reh)
}
\arguments{
\item{edgelist}{input edgelist from 'reh' object}
}
\value{
data.frame
}
\description{
convertToReleventEdgelist()
}
