% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decay}
\alias{decay}
\title{decay (function that calculates the decay according)}
\usage{
decay(type, x, pars)
}
\arguments{
\item{type}{string defining the type of decay}

\item{x}{elapsed time of interest}

\item{pars}{is the vector with order}
}
\description{
decay (function that calculates the decay according)
}
