% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMA.R
\name{DoAnalysis}
\alias{DoAnalysis}
\title{DoAnalysis}
\usage{
DoAnalysis(
  env = NULL,
  do = c("mle", "sd", "sd_increasing_time_lengths", "mle_precision_fit", "mle_width",
    "mle_elapsed_time", "weights_precision", "fit_precision_interval",
    "mle_transparent_lines", "beta_loglik", "widths_weights"),
  weights = c("BIC", "pseudoBMAplus", "stacking", "WAIC", "loglik", "all"),
  K = 2,
  n_knots = 300,
  top = 10
)
}
\arguments{
\item{env}{environment where to work}

\item{do}{which analysis to perform}

\item{weights}{which vector of weights to consider}

\item{K}{which dimension}

\item{n_knots}{number of knots}

\item{top}{best top-x models to select}
}
\value{
ggplot objects depending on the specified type
}
\description{
A function which plots posterior trends and eventually returns the ggplot objects
}
