% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{drawFromPosterior}
\alias{drawFromPosterior}
\title{drawFromPosterior}
\usage{
drawFromPosterior(input, weights, nsim, elapsed_time, max_width)
}
\arguments{
\item{input}{stepwiseModelsREH environment}

\item{weights}{list of vectors of normalized weights (BIC, pseudoBMA+, ...)}

\item{nsim}{number of simulations from the posterior}

\item{elapsed_time}{number of knots for the elapsed time axis}

\item{max_width}{maximum widths of intervals (used in the "stepwise_models" method)}
}
\value{
posterior estimates of memory decay per each type of weights
}
\description{
A function which return a list of draws according to different possible weighting systems
}
