% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpd.R
\name{elpd}
\alias{elpd}
\title{elpd (Expected Log-pointwise Predictive Density)}
\usage{
elpd(input)
}
\arguments{
\item{input}{input for elpd function}
}
\value{
vector of elpd
}
\description{
A function which return a list of weights according to different possible weighting systems
}
