% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{formula_lab}
\alias{formula_lab}
\title{formula_lab}
\usage{
formula_lab(input)
}
\arguments{
\item{input}{input of the function is a}
}
\value{
the function returns a NULL
}
\description{
A function where do build up a formula object for smm and pmm
}
