% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateREH.R
\name{generateREH}
\alias{generateREH}
\title{generateREH}
\usage{
generateREH(
  N,
  max_time = NULL,
  M,
  exo_stats = NULL,
  beta_endo,
  beta_exo = NULL,
  seed = NULL
)
}
\arguments{
\item{N}{number of actors in the network.}

\item{max_time}{this parameter is not used in this function anymore. By setting the maximum value for the variable time we cannot define the number of final events in advance.}

\item{M}{number of events we want to generate.}

\item{exo_stats}{matrix of N*(N-1) rows with as many columns as the number of exogenous statistics.}

\item{beta_endo}{named list of endogenous statistics with decay type (decay_type) and parameters (pars). Names must match with the ones used within the function (see vector of names).}

\item{beta_exo}{vector of effects (its length must be equal to the number of columns of exo_stats).}

\item{seed}{used internally. If not specified a random seed is used and finally returned.}
}
\value{
list of final dgelist, seed used to generate it and vector of offset values.
}
\description{
A function which generate a relational event history according to network characteristics and statistics that are specified by input arguments
}
