% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBinaryREH}
\alias{getBinaryREH}
\title{getBinaryREH}
\usage{
getBinaryREH(dyad, D)
}
\arguments{
\item{dyad}{vector of dyad occurred (reh$edgelist[,2])}

\item{D}{numbr of possible dyads (reh$D)}
}
\value{
matrix with dimensions given by the number of events (by row) and the dimension of the risk set (by column) .
}
\description{
getBinaryREH
}
