% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCounts}
\alias{getCounts}
\title{getCounts}
\usage{
getCounts(dyad, D, lbs_ubs, ncores)
}
\arguments{
\item{dyad}{is the vector of observed dyads}

\item{D}{is the number of dyads}

\item{lbs_ubs}{matrix of lower bounds (lb's, first column) and upper bounds (ub's, second column) of binaryREH}

\item{ncores}{ncores}
}
\value{
matrix of accumulated counts within ranges of binaryREH
}
\description{
getCounts
}
