% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCountsIndex}
\alias{getCountsIndex}
\title{getCountsIndex (function useful for statistics computation)}
\usage{
getCountsIndex(intervals, counts, ncores)
}
\arguments{
\item{intervals}{matrix of intervals (non-unique intervals, first two columns are lower bound and upper bound respectively)}

\item{counts}{matrix of intervals (first two columns are lower bound and upper bound respectively)}

\item{ncores}{number of ncores}
}
\value{
vector of row indices relative to the matrix 'counts'
}
\description{
getCountsIndex (function useful for statistics computation)
}
