% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCountsIndex_old}
\alias{getCountsIndex_old}
\title{getCountsIndex_old (function useful for statistics computation)}
\usage{
getCountsIndex_old(intervals, counts)
}
\arguments{
\item{intervals}{matrix of intervals (non-unique intervals, first two columns are lower bound and upper bound respectively)}

\item{counts}{matrix of intervals (first two columns are lower bound and upper bound respectively)}
}
\value{
vector of row indices relative to the matrix 'counts'
}
\description{
getCountsIndex_old (function useful for statistics computation)
}
