% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCountsOMP_old}
\alias{getCountsOMP_old}
\title{getCountsOMP_old}
\usage{
getCountsOMP_old(binaryREH, lbs_ubs, ncores)
}
\arguments{
\item{binaryREH}{matrix}

\item{lbs_ubs}{matrix of lower bounds (lb's, first column) and upper bounds (ub's, second column) of binaryREH}

\item{ncores}{ncores}
}
\value{
matrix of accumulated counts within ranges of binaryREH
}
\description{
getCountsOMP_old
}
