% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDraws_old}
\alias{getDraws_old}
\title{getDraws_old}
\usage{
getDraws_old(sample_models, which_pars, n_pars, n_stats, input, knots_seq)
}
\arguments{
\item{sample_models}{models sample with probabilities given by the specific weighting system used (BIC, pseudoBMA+, stacking, WAIC)}

\item{which_pars}{which parameters}

\item{n_pars}{number of parameters per model in the set}

\item{n_stats}{number of statistics}

\item{input}{list() of K (number of intervals), gammas (time widths), mle_betas (Maximum Likelihood Estimates), vcov_betas (variance and covariance matrix)}

\item{knots_seq}{sequence of time widths at which compute the posterior distribution}
}
\value{
matrix.
}
\description{
Function to draw from the posterior distribution (here the second step of drawing from Normal distribution is performed;
the first consists in drawing from the multinomial distribution for the Q models)
}
