% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEndoEffects_old}
\alias{getEndoEffects_old}
\title{getEndoEffects_old}
\usage{
getEndoEffects_old(env)
}
\arguments{
\item{env}{environment where the user is currently working (usually the global one which can be accessed via 'globalenv()' function or '.GlobalEnv' object)}
}
\value{
array of Statistics specified in the
}
\description{
getEndoEffects_old
}
