% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIntervals}
\alias{getIntervals}
\title{getIntervals}
\usage{
getIntervals(env, widths, time, M, K_q)
}
\arguments{
\item{env}{is the global environment}

\item{widths}{time lengths of intervals as to the stepwise model}

\item{time}{vector of time points}

\item{M}{number of events}

\item{K_q}{number of steps for the q-th model}
}
\value{
matrix.
}
\description{
given vectors of widths and time, the function returns.
}
