% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIntervals_old}
\alias{getIntervals_old}
\title{getIntervals_old}
\usage{
getIntervals_old(env)
}
\arguments{
\item{env}{is the global environment}
}
\value{
matrix.
}
\description{
given vectors of widths and time, the function returns.
}
