% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIntervalStatistics}
\alias{getIntervalStatistics}
\title{getIntervalStatistics}
\usage{
getIntervalStatistics(
  env,
  counts,
  intervals,
  actor1,
  actor2,
  time,
  M,
  N,
  D,
  K,
  statistics,
  model,
  senderRate = FALSE,
  ncores = 1L
)
}
\arguments{
\item{counts}{matrix of dyad counts for unique intervals}

\item{intervals}{matrix with all intervals}

\item{actor1}{time-ordered sequence of actor1 ID's}

\item{actor2}{time-ordered sequence of actor2 ID's}

\item{time}{vector of time}

\item{M}{number of events}

\item{N}{number of actors}

\item{D}{number of dyads (no event type)}

\item{K}{number of intervals (q-th model)}

\item{statistics}{string vector of endogenous statistics}

\item{model}{'tie' or 'actor'}

\item{senderRate}{true/false for actor-oriented modeling. if true, then statistics for sender-rate model, otherwise statistics for the receiver-choice model}

\item{ncores}{number of ncores to create in case of parallelization}

\item{stats}{is a list of statistics (exogenous and endogenous) which will be updated with the endogenous statistics}
}
\value{
list of (interval) statistics
}
\description{
getIntervalStatistics
}
