% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMA.R
\name{getMLEreh}
\alias{getMLEreh}
\title{getMLEreh}
\usage{
getMLEreh(edgelist, exo_stats, par_vec, n_cores)
}
\arguments{
\item{edgelist}{edgelist to use for the estimation}

\item{exo_stats}{matrix of exogenous statistics}

\item{par_vec}{vector of values for the halflife/tmax}

\item{n_cores}{number of cores to use in the parallelization}
}
\value{
list of two objects: "halflife" containing vectors with parameters estimates give by exponential decay (halflife); "tmax", like "halflife"but assuming a one step decay.
}
\description{
A function which returns estimates
}
