% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMA.R
\name{getMLEsims}
\alias{getMLEsims}
\title{getMLEsims}
\usage{
getMLEsims(list_edgelist, exo_stats, par_vec, n_cores, file_name)
}
\arguments{
\item{list_edgelist}{list of edgelists to use for the estimation}

\item{exo_stats}{matrix of exogenous statistics}

\item{par_vec}{vector of values for the halflife/tmax}

\item{n_cores}{number of cores to use in the parallelization}

\item{file_name}{string with name to use for the partial savings}
}
\value{
list of two objects: "halflife" containing vectors with parameters estimates give by exponential decay (halflife); "tmax", like "halflife"but assuming a one step decay.
}
\description{
A function which runs simulations
}
