% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{getModelsWeights}
\alias{getModelsWeights}
\title{getModelsWeights}
\usage{
getModelsWeights(
  input = NULL,
  weights = c("BIC", "pseudoBMAplus", "stacking", "WAIC", "loglik", "elpd_lfo", "all")
)
}
\arguments{
\item{input}{list of vector, matrices and array useful for weights computation}

\item{weights}{type of weighting system to be used}
}
\value{
list of normalized weights
}
\description{
A function which return a list of weights according to different possible weighting systems
}
