% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMA.R
\name{getPosterior}
\alias{getPosterior}
\title{getPosterior}
\usage{
getPosterior(
  env = NULL,
  weights = c("BIC", "pseudoBMAplus", "stacking", "WAIC", "loglik", "elpd_lfo", "all"),
  method = c("default", "spline"),
  n_knots = 300,
  nsim = 5000,
  B = 500,
  robustify = FALSE,
  R = 100,
  subset_K = NULL,
  sample_size_per_K = 10,
  plot = FALSE,
  top_K_models = 10
)
}
\arguments{
\item{env}{environment stepwiseModelsREH where models are stores}

\item{weights}{which weighting system to use}

\item{method}{estimation method to use in the bootstrap}

\item{n_knots}{number of knots of the elapsed time where to evaluate the posterior distribution}

\item{nsim}{number of posterior draws}

\item{B}{number of splines}

\item{robustify}{TRUE/FALSE whether to get more robust estimate by emulating R times a BMA}

\item{R}{number of BMAs replications, when the argument robustify is TRUE}

\item{subset_K}{if user wants specifically to subset the analysis to specific interval dimensions}

\item{sample_size_per_K}{number of models to sample per K (number of intervals), when robustify is TRUE: stratified sampling}

\item{plot}{TRUE/FALSE plotting results}

\item{top_K_models}{find the top K models for each weighting system and return: indices, widths and mles (inside a list)}
}
\value{
list of if robustify FALSE (normalized weights, posterior distributions across gamma's and per each weight distribution); if robustify is TRUE ...
}
\description{
A function which return a list of weights according to different possible weighting systems
}
