% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMA.R
\name{getPosteriorDefault}
\alias{getPosteriorDefault}
\title{getPosteriorDefault}
\usage{
getPosteriorDefault(
  env = NULL,
  weights = c("BIC", "pseudoBMAplus", "stacking", "WAIC", "loglik", "all"),
  nsim = 10000,
  n_knots = 1000
)
}
\arguments{
\item{env}{environment/list stepwiseModelsREH where models are stored (or a subset of it)}

\item{weights}{type of weighting system to be used}

\item{nsim}{number of simulations per gamma}

\item{n_knots}{number of gamma's}
}
\value{
list of: normalized weights, posterior distributions across gamma's and per each weight distribution
}
\description{
A function which return a list of weights according to different possible weighting systems
}
