% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{getPosteriorSpline}
\alias{getPosteriorSpline}
\title{getPosteriorSpline}
\usage{
getPosteriorSpline(
  env = NULL,
  weights = c("BIC", "pseudoBMAplus", "stacking", "WAIC", "loglik", "elpd_lfo", "all"),
  nsim = 10000,
  B = 1000,
  n_knots = 1000
)
}
\arguments{
\item{env}{environment/list stepwiseModelsREH where models are stored (or a subset of it)}

\item{weights}{type of weighting system to be used}

\item{nsim}{number of simulations per gamma}

\item{B}{number of splines}

\item{n_knots}{number of gamma's}
}
\value{
list of: normalized weights, posterior distributions across gamma's and per each weight distribution
}
\description{
A function which return....
}
