% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRiskset}
\alias{getRiskset}
\title{getRiskset (obtain permutations of actors' id).}
\usage{
getRiskset(actors_id, N, selfedges)
}
\arguments{
\item{N}{number of actors in the network.}

\item{selfedges}{if selfedges are allowed (TRUE) or not (FALSE).}

\item{actors}{vector of actors' id's (either integer or char vector).}
}
\value{
matrix of possible dyadic events.
}
\description{
getRiskset (obtain permutations of actors' id).
}
