% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRiskset_old}
\alias{getRiskset_old}
\title{getRiskset_old (obtain permutations of actors' id). (will become getRisksetMatrix)}
\usage{
getRiskset_old(actors_id, N, selfedges)
}
\arguments{
\item{N}{number of actors in the network.}

\item{selfedges}{if selfedges are allowed (TRUE) or not (FALSE).}

\item{actors}{vector of actors' id's (either integer or char vector).}
}
\value{
matrix of possible dyadic events.
}
\description{
getRiskset_old (obtain permutations of actors' id). (will become getRisksetMatrix)
}
