% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRisksetMatrix_old}
\alias{getRisksetMatrix_old}
\title{getRisksetMatrix_old (will become getRisksetCube) [[ TO REMOVE ]]}
\usage{
getRisksetMatrix_old(riskset, N)
}
\arguments{
\item{riskset}{matrix (sender and receiver integer id's by column)}

\item{N}{number of actors in the network.}
}
\value{
matrix of possible dyadic events arranged in a square matrx (cell value is the row index of the riskset matrix input, which will be the same to the column index in binary REH).
}
\description{
getRisksetMatrix_old (will become getRisksetCube) [[ TO REMOVE ]]
}
