% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSmoothEndoEffects}
\alias{getSmoothEndoEffects}
\title{getSmoothEndoEffects}
\usage{
getSmoothEndoEffects(reh, endo_effects, endo_memory_pars, n_threads)
}
\arguments{
\item{endo_effects}{string vector indicating the endogenous effects}

\item{n_threads}{number of corse to be used in the parallelization}

\item{env}{environment where the user is currently working (usually the global one which can be accessed via 'globalenv()' function or '.GlobalEnv' object)}

\item{decay_fun}{function for the decay of past events influence}

\item{pars}{matrix with parameters for the decay_fun}
}
\value{
array of Statistics specified
}
\description{
getSmoothEndoEffects
}
