% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStepwiseModels.R
\name{getStepwiseModels}
\alias{getStepwiseModels}
\title{getStepwiseModels}
\usage{
getStepwiseModels(
  env = globalenv(),
  effects = NULL,
  K_range = c(2:6),
  max_width = NULL,
  nsim_per_K = 200,
  L = NULL,
  A_SAP = 1,
  file_name = "simulation_0",
  WAIC = TRUE,
  n_cores = 10
)
}
\arguments{
\item{env}{environment where the user is currently working}

\item{effects}{list of two objects: effects$endogenous is a vector of strings with specified the effects of interest; effects$exogenous is an array [M*dyads*exogenous_effects]}

\item{K_range}{range as to the number of intervals that are going to be generated.}

\item{max_width}{maximum for width}

\item{nsim_per_K}{how many simulations per number of intervals}

\item{L}{number of events to use as history to estimate the predictive pointwise density of future events (set to M-10 by default)}

\item{A_SAP}{A-Steps-Ahead-Predictions is the number of steps ahead we want consider in the prediction (set to 1 by default)}

\item{file_name}{a string indicating the name of the file .RData that will save results each 20 models.}

\item{WAIC}{default is TRUE and is needed where WAIC weights have to be estimated}

\item{n_cores}{number of threads to create}
}
\value{
the function doesn't return any output but updates objects inside the environment 'stepwiseModelsREH'
}
\description{
A function which run the estimation of Q models and updates the environment 'stepwiseModelsREH'
}
