% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStepwiseModelsOLD.R
\name{getStepwiseModelsOLD}
\alias{getStepwiseModelsOLD}
\title{getStepwiseModelsOLD}
\usage{
getStepwiseModelsOLD(
  env = globalenv(),
  effects = NULL,
  K = c(2:6),
  max_width = NULL,
  min_diff = NULL,
  nsim_per_K = 101,
  L = NULL,
  A_SAP = 1,
  file_name = "simulation_0",
  WAIC = TRUE,
  n_cores = 10,
  first_event = NULL,
  last_event = NULL,
  widths_input = NULL,
  K_input = NULL,
  interval_type_input = NULL
)
}
\arguments{
\item{env}{environment where the user is currently working}

\item{effects}{list of two objects: effects$endogenous is a vector of strings with specified the effects of interest; effects$exogenous is an array [M*dyads*exogenous_effects]}

\item{K}{range as to the number of intervals that are going to be generated.}

\item{max_width}{maximum for width}

\item{min_diff}{minimum width (diff will be changed with width)}

\item{nsim_per_K}{how many simulations per number of intervals}

\item{L}{number of events to use as history to estimate the predictive pointwise density of future events (set to M-10 by default)}

\item{A_SAP}{A-Steps-Ahead-Predictions is the number of steps ahead we want consider in the prediction (set to 1 by default)}

\item{file_name}{a string indicating the name of the file .RData that will save results each 20 models.}

\item{WAIC}{default is TRUE and is needed where WAIC weights have to be estimated}

\item{n_cores}{number of threads to create}

\item{first_event}{integer indicating the event at which start the event sequence and statistic in the estimation stage}

\item{last_event}{integer indicating the event at which stop the event sequence and statistic in the estimation stage}

\item{widths_input}{matirx of sequences of widths from the user}

\item{K_input}{vector of number of steps per each sequence of widths from the user}

\item{interval_type_input}{vector of interval types per each sequence of widths from the user}
}
\value{
the function doesn't return any output but updates objects inside the environment 'stepwiseModelsREH'
}
\description{
A function which run the estimation of Q models and updates the environment 'stepwiseModelsREH'
}
