% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getWAIC_old}
\alias{getWAIC_old}
\title{getWAIC_old}
\usage{
getWAIC_old(pars, stats, events, interevent_time)
}
\arguments{
\item{pars}{parameters values}

\item{stats}{sub-array of statistics}

\item{events}{sub-matirx of events}

\item{interevent_time}{sub-vector of interevent times}
}
\value{
matrix of p_m and lpd_m values for a specific stepwise model
}
\description{
getWAIC_old
}
