% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{halflifeWeight}
\alias{halflifeWeight}
\title{halflifeWeight}
\usage{
halflifeWeight(x, pars)
}
\arguments{
\item{x}{transpired time}

\item{pars}{is a column vector with the halflife}
}
\description{
halflifeWeight
}
