% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeREH.R
\name{initializeREH}
\alias{initializeREH}
\title{initializeREH (preprocessing raw data in a suitable format to the 'bremory' package, I used this as starting function for the reh() in remstimate)}
\usage{
initializeREH(
  edgelist,
  riskset = NULL,
  directed = TRUE,
  weights = NULL,
  env = globalenv()
)
}
\arguments{
\item{edgelist}{a matrix of dimensions (number of events * 3). By column (time,sender,receiver) of relational events. The first column is the time: if the timing is "ordinal" then a vector of increasing integer (from 1 to the number of events) is provided. Sender and receivers can be provided as either char or integer type.}

\item{riskset}{a matrix of dimensions [number of dyads in the risk set * 2]. By column (sender,receiver) of the dyad in the risk set. Sender and receivers can be provided as either char or integer type. The risk set must contain only possible interactions between actors, possbile event types must be excluded in the definition of the risk set.}

\item{directed}{it is set TRUE by default. If FALSE, statistics where directionality of dyads matters will be summed up: e.g., inertia(i,j)+inertia(j,i).}

\item{env}{environment where the user wants to store the preprocessed output (it is set to .GlobalEnv by default).}

\item{eventType}{vector of event types.}
}
\value{
The environment in which the user prefers to work is updated by including vectors, matrices and other data structures necessary to pursue the analysis.
}
\description{
initializeREH (preprocessing raw data in a suitable format to the 'bremory' package, I used this as starting function for the reh() in remstimate)
}
