% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\title{intervals}
\usage{
intervals(
  K = NULL,
  nsimK = 100,
  maxWidth = NULL,
  minDiff = NULL,
  intervals = c("all", "increasing", "decreasing", "equal")
)
}
\arguments{
\item{K}{intervals dimensions}

\item{nsimK}{number of simulations per K}

\item{maxWidth}{maximum time length}

\item{minDiff}{minimum width (diff will be changed to width)}

\item{intervals}{type of intervals to be generated: "all" it will return an object of class "intervals" containing nsimK intervals increasing, nsimK intervals decreasing and 1 interval equal per each K; if "increasing" or "decreasing" only nsimK increasing intervals per each K will be generated; if "equal" only one equal size interval will be returned per each K.}
}
\value{
the function updates the object width in the getStepwiseModels environment.
}
\description{
A function which return an array with all the statistics, both exogenous and endogenous.
}
