% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linearWeight}
\alias{linearWeight}
\title{linearWeight}
\usage{
linearWeight(x, pars)
}
\arguments{
\item{x}{transpired time}

\item{pars}{is the vector with order (alpha,psi) with alpha in (0,Inf) and psi in (0,Inf)}
}
\description{
linearWeight
}
