% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logpJi}
\alias{logpJi}
\title{logpJi}
\usage{
logpJi(pars, stats, event, interevent_time)
}
\arguments{
\item{pars}{is a matrix of parameters (note: the order must be aligned witht the column order in 'stats')}

\item{stats}{is a matrix of dimensions n_dyads*variables with statistics of interest by column and dyads by row.}

\item{event}{is a vector of 1/0 : 1 indicating the observed dyad and 0 the non observed dyads.}

\item{interevent_time}{the time difference between the current time point and the previous event time.}
}
\value{
vector of log-pointwise density values of a specific time point at different parameters values
}
\description{
logpJi
}
