% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{merge.smm}
\alias{merge.smm}
\title{merge.smm}
\usage{
\method{merge}{smm}(x)
}
\arguments{
\item{x}{a list of \code{smm} objects}
}
\description{
A function that given a list of smm object, it merges into a single 'smm' object
}
