% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nllik}
\alias{nllik}
\title{nllik (negative log-likelihood rem model)}
\usage{
nllik(pars, stats, event_binary, interevent_time, n_cores)
}
\arguments{
\item{pars}{is a vector of parameters (note: the order must be aligned witht the column order in 'stats')}

\item{stats}{is a cube of dimensions n_dyads*variables*M with statistics of interest by column and dyads by row.}

\item{event_binary}{is a matrix of ones and zeros of dimensions M*n_dyads : 1 indicating the observed dyad and 0 the non observed dyads.}

\item{interevent_time}{the vector of time differences between the current time point and the previous event time.}

\item{n_cores}{n_cores}
}
\value{
log-pointwise density value of a specific time point
}
\description{
nllik (negative log-likelihood rem model)
}
