% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{performBSIR}
\alias{performBSIR}
\title{performBSIR}
\usage{
performBSIR(
  nsim,
  mean,
  sigma,
  df,
  stats,
  event_binary,
  interevent_time,
  nthreads
)
}
\arguments{
\item{nsim}{number of simulations}

\item{mean}{vector of model MLEs}

\item{sigma}{matric of variances and covariances}

\item{df}{degrees of freedom for the multivariate Student t (used as proposal distribution)}

\item{stats}{cube of statistics [D*P*M]}

\item{event_binary}{matrix of 1/0}

\item{interevent_time}{vector of interevent times}

\item{nthreads}{number of cores to be used in the parallelized calculation of the nllik()}
}
\value{
log-pointwise density value of a specific time point
}
\description{
A function that performs the BSIR on the REM model
}
