% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMA.R
\name{plot.bma}
\alias{plot.bma}
\title{plot.bma}
\usage{
\method{plot}{bma}(
  bma,
  ci = FALSE,
  xlim,
  width,
  height,
  ncol,
  nrow,
  save = TRUE,
  widthMax = NULL,
  output_folder_name = "output_plots"
)
}
\arguments{
\item{bma}{is a \code{bma} object .. draws}

\item{ci}{...}

\item{xlim}{...}

\item{width}{...}

\item{height}{...}

\item{ncol}{...}

\item{nrow}{...}

\item{save}{...}

\item{widthMax}{...gamma_max}

\item{output_folder_name}{...}

\item{ylim}{...}

\item{true_pars}{... list of ...}
}
\description{
A function that returns a summary of the temporal network.
}
