% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMA.R
\name{plotPosterior}
\alias{plotPosterior}
\title{plotPosterior}
\usage{
plotPosterior(
  posterior_draws,
  type = c("trend", "credibility intervals"),
  true_effects = NULL,
  save = TRUE,
  file_name = NULL
)
}
\arguments{
\item{posterior_draws}{list of posterior draws from different weighting systems (BIC, pseudoBMA+, ...)}

\item{type}{type of plot: trend type will just plot the median, credibility intervals type will also plot the credibitlity interval at 0.95}

\item{true_effects}{effects list of endo effects and exo_effects}

\item{save}{TRUE/FALSE whether to plot results right after the creation of the list of ggplot objects}

\item{file_name}{file name is usually a combination of method and subset_K parameter from the getPosterior function (it is useful just for now in order to save results fastly)}
}
\value{
ggplot objects depending on the specified type
}
\description{
A function which plots posterior trends and eventually returns the ggplot objects
}
