% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm.R
\name{pmm}
\alias{pmm}
\title{Parametric Memory Model}
\usage{
pmm(formula, reh, memory, data = NULL, ncores = 1L, ...)
}
\arguments{
\item{formula}{formula specifing the linear predictor (list of formulas if actor-oriented model)}

\item{reh}{either a 'reh' object (see remify::remify() function) or the edgelist to process}

\item{memory}{memory shape to use, if parameter is specified then no optimization is carried out}

\item{data}{[\emph{optional}] list of exogenous variables (time varying or not)}

\item{ncores}{[\emph{optional}] number of ncores to use for the parallelization}

\item{...}{[\emph{optional}] other arguments referring to those in bremory::elpd()}
}
\value{
object of class 'pmm' with attributes and methods explained in the vignette(topic = "pmm", package = "bremory")
}
\description{
A function for the estimation of one or more step-wise models (at the moment working without remstats package)
}
\examples{

# examples here
}
