% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStepwiseModels.R
\name{PSIS_LFO_CV_1_SAP}
\alias{PSIS_LFO_CV_1_SAP}
\title{PSIS_LFO_CV_1_SAP}
\usage{
PSIS_LFO_CV_1_SAP(
  n_dyads,
  eventlist,
  statslist,
  t,
  first_event,
  last_event,
  L,
  n_sim_is,
  tau,
  cores
)
}
\arguments{
\item{n_dyads}{...}

\item{eventlist}{...}

\item{statslist}{...}

\item{t}{...}

\item{first_event}{...}

\item{last_event}{...}

\item{L}{...}

\item{n_sim_is}{...}

\item{tau}{...}

\item{cores}{...}
}
\value{
the function doesn't return any output but updates objects inside the environment 'stepwiseModelsREH'
}
\description{
A function which run th PSIS with LFO CV by using 1-Step-Ahead-Predictions
}
