% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{smm}
\alias{smm}
\title{smm (Stepwise Memory Models)}
\usage{
smm(
  formula = NULL,
  reh = NULL,
  data = NULL,
  effects = NULL,
  intervals = NULL,
  WAIC = TRUE,
  ELPD = FALSE,
  firstEvent = NULL,
  lastEvent = NULL,
  nthreads = NULL,
  filename = "simulation_0",
  env = globalenv(),
  ...
)
}
\arguments{
\item{formula}{[not working]formula specifing the linear predictor (this argument is still being coded, plase use the argument 'effects')}

\item{reh}{either a 'reh' object (see remify::reh() function) or the edgelist}

\item{data}{[not working]list of exogenous variables (time varying or not) (this argument is still being coded, please use the argument 'effects')}

\item{effects}{[working]list of two objects: effects$endogenous is a vector of strings with specified the effects of interest; effects$exogenous is an array [M*dyads*exogenous_effects]}

\item{intervals}{it can be specified in two different ways: (1) as a list of intervals with "widths", "K"; (2) as list specifying: a vector of integers indicating the number of steps per stepwise model ("K"), the maximum time width for intervals ("maxWidth"), the minimum width for intervals ("minDiff"), the number of simulations per each K ("nsimK"). With the second specification, the function bremory::intervals() will be used.}

\item{WAIC}{default is TRUE and is needed where WAIC weights have to be estimated}

\item{ELPD}{default is FALSE and is TRUE if ELPD-LFO has to be calculated (PSIS approximation is used [ref. to literature] for one-step-ahead predictions). This procedure might slow down the algorithm.}

\item{firstEvent}{integer indicating the event at which start the event sequence and statistic in the estimation stage}

\item{lastEvent}{integer indicating the event at which stop the event sequence and statistic in the estimation stage}

\item{nthreads}{number of threads to create}

\item{filename}{a string indicating the name of the file .RData that will save results each 20 models.}

\item{env}{environment where the user is currently working (default is set to .GlobalEnv, globalenv())}

\item{...}{other arguments referring to those in bremory::elpd()}
}
\value{
object of class 'ssm'
}
\description{
A function which run the estimation of Q models and updates the environment 'stepwiseModels'
}
