% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{smmTemp}
\alias{smmTemp}
\title{smmTemp (Stepwise Memory Models [Temporary Function])}
\usage{
smmTemp(
  formula,
  reh,
  data,
  subset,
  intervals = NULL,
  WAIC = TRUE,
  ELPD = FALSE,
  threads = parallel::detectCores() - 2,
  env = globalenv(),
  ...
)
}
\arguments{
\item{formula}{formula specifing the linear predictor (this argument is still being coded, plase use the argument 'effects')}

\item{reh}{either a 'reh' object (see remify::reh() function) or the edgelist}

\item{data}{list of exogenous variables (time varying or not) (this argument is still being coded, please use the argument 'effects')}

\item{subset}{old {firstEvent, lastEvent}}

\item{intervals}{it can be specified in two different ways: (1) as a list of intervals with "widths", "K" and "type"; (2) as list specifying: a vector of integers indicating the number of steps per stepwise model ("K"), the maximum time width for intervals ("maxWidth"), the minimum width for intervals ("minDiff"), the number of simulations per each K ("nsimK"). With the second specification, the function bremory::intervals() will be used.}

\item{WAIC}{default is TRUE and is needed where WAIC weights have to be estimated}

\item{ELPD}{default is FALSE and is TRUE if ELPD-LFO has to be calculated (PSIS approximation is used [ref. to literature] for one-step-ahead predictions). This procedure might slow down the algorithm.}

\item{threads}{number of threads to create}

\item{env}{environment where the user is currently working (default is set to .GlobalEnv, globalenv())}

\item{...}{other arguments referring to those in bremory::elpd()}
}
\value{
object of class 'ssm'
}
\description{
A function which run the estimation of Q models and updates the environment 'stepwiseModels'
}
