% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smoothing_posterior}
\alias{smoothing_posterior}
\title{smoothing posterior.}
\usage{
smoothing_posterior(
  sample_models,
  knots_seq,
  which_pars,
  post_betas,
  post_gammas,
  U
)
}
\arguments{
\item{sample_models}{sample_models}

\item{knots_seq}{knots_seq}

\item{which_pars}{which_pars}

\item{post_betas}{post_betas}

\item{post_gammas}{post_gammas}

\item{n_pars}{n_pars}
}
\value{
a cube of posterior draws.
}
\description{
This function calculates the posterior distribution give...
}
