% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticsREH.R
\name{statisticsREH}
\alias{statisticsREH}
\title{statisticsREH}
\usage{
statisticsREH(reh = NULL, effects = NULL, env = globalenv())
}
\arguments{
\item{reh}{'reh' object}

\item{effects}{list of two objects: effects$endogenous is a vector of strings with specified the effects of interest; effects$exogenous is a matrix [dyads*exogenous_effects]}

\item{env}{environment where the user is currently working}
}
\value{
size of the specified environment.
}
\description{
A function which creates an environment to...
}
