% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tmaxWeight}
\alias{tmaxWeight}
\title{tmaxWeight}
\usage{
tmaxWeight(x, pars)
}
\arguments{
\item{x}{transpired time}

\item{pars}{is the matrix of dim 1x1 with the time threshold}
}
\description{
tmaxWeight
}
