% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tryClone}
\alias{tryClone}
\title{tryClone}
\usage{
tryClone(lambda, index, stats_col)
}
\arguments{
\item{lambda}{aa}

\item{index}{bb}

\item{stats_col}{cc}
}
\value{
input [dataframe]
}
\description{
tryClone
}
