% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{twoStepsWeight}
\alias{twoStepsWeight}
\title{twoStepsWeight}
\usage{
twoStepsWeight(x, pars)
}
\arguments{
\item{x}{transpired time}

\item{pars}{is the vector of parameters OLD=(shape,scale,psi,phi,shape,scale), (shape1,scale1,psi,phi,x0,shape2,scale2)}
}
\description{
twoStepsWeight
}
