% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateEndoEffects}
\alias{updateEndoEffects}
\title{updateEndoEffects}
\usage{
updateEndoEffects(
  effects,
  beta_endo,
  binaryREH,
  elapsed_time,
  edgelist,
  riskset,
  riskset_matrix,
  N,
  M_partial
)
}
\arguments{
\item{effects}{string vector of endogenous effects to calculate}

\item{beta_endo}{list of parameters per each effect (the order must follow the one of the vector 'effects')}

\item{binaryREH}{binary Relational Event History until event at t(M_partial)}

\item{elapsed_time}{vector of elapsed times since each past event until t(M_partial)}

\item{edgelist}{edgelist (partial edgelist until M_partial)}

\item{riskset}{matrix of (sender,receiver) combinations in the case of no selfedges}

\item{riskset_matrix}{square matrix where the (row,colum)=(sender,receiver) returns the corresponding column index in the matrix binaryREH_partial}

\item{N}{number of actors}

\item{M_partial}{partial event sequence length}
}
\value{
matrix of statistics per dyad (where a continuous effect decay for each statistics was specified)
}
\description{
updateEndoEffects
}
