% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weibullWeight}
\alias{weibullWeight}
\title{weibullWeight (it can be either exponential of a smoothed single-step decrease)}
\usage{
weibullWeight(x, pars)
}
\arguments{
\item{x}{transpired time}

\item{pars}{is the vector with order (shape, scale, psi)}
}
\description{
weibullWeight (it can be either exponential of a smoothed single-step decrease)
}
