% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWidths.R
\name{generateWidths}
\alias{generateWidths}
\title{generateWidths}
\usage{
generateWidths(
  env = globalenv(),
  K_range = c(2:6),
  nsim_per_K = 200,
  max_width = NULL
)
}
\arguments{
\item{env}{environment where the user is currently working}

\item{K_range}{intervals dimensions}

\item{nsim_per_K}{number of simulations per K}

\item{max_width}{maximum time length}
}
\value{
the function updates the object width in the getStepwiseModels environment.
}
\description{
A function which return an array with all the statistics, both exogenous and endogenous.
}
