% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBinaryREH}
\alias{getBinaryREH}
\title{getBinaryREH()}
\usage{
getBinaryREH(M, N, edgelist, riskset_matrix)
}
\arguments{
\item{M}{number of relational events observed}

\item{N}{number of actors}

\item{edgelist}{matrix of (time,sender,receiver)}

\item{riskset_matrix}{matrix of possible dyadic events with column_order per each cell (this will be the order in the binaryREH matrices columns)}
}
\value{
Matrix with dimensions given by the number of events (by row) and the dimension of the risk set (by column) .
}
\description{
getBinaryREH()
}
