% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCountsOMP}
\alias{getCountsOMP}
\title{getCountsOMP}
\usage{
getCountsOMP(binaryREH, lbs_ubs, n_cores)
}
\arguments{
\item{binaryREH}{matrix}

\item{lbs_ubs}{matrix of lower bounds (lb's, first column) and upper bounds (ub's, second column) of binaryREH}

\item{n_cores}{n_cores}
}
\value{
matrix of accumulated counts within ranges of binaryREH
}
\description{
getCountsOMP
}
