% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIntervals}
\alias{getIntervals}
\title{getIntervals}
\usage{
getIntervals(env, widths)
}
\arguments{
\item{env}{is the global environment}

\item{widths}{time lengths of intervals as to the stepwise model}
}
\value{
matrix.
}
\description{
given vectors of widths and time, the function returns.
}
