% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSmoothEndoEffects}
\alias{getSmoothEndoEffects}
\title{getSmoothEndoEffects}
\usage{
getSmoothEndoEffects(reh, endo_effects, endo_memory_pars, nthreads)
}
\arguments{
\item{reh}{"reh" object}

\item{endo_effects}{string vector indicating the endogenous effects}

\item{endo_memory_pars}{function for the decay of past events influence}

\item{nthreads}{number of corse to be used in the parallelization}
}
\value{
array of Statistics specified
}
\description{
getSmoothEndoEffects
}
