% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{merge.smm}
\alias{merge.smm}
\title{merge.smm}
\usage{
\method{merge}{smm}(smm)
}
\arguments{
\item{smm}{a list of \code{smm} objects}
}
\description{
A function that returns a summary of the temporal network.
}
