% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{smm}
\alias{smm}
\title{Step-wise Memory Model}
\usage{
smm(
  formula,
  reh,
  intervals,
  data = NULL,
  WAIC = TRUE,
  ELPD = FALSE,
  ncores = 1L,
  subset = NULL,
  ...
)
}
\arguments{
\item{formula}{formula specifing the linear predictor (list of formulas if actor-oriented model)}

\item{reh}{either a 'reh' object (see remify::remify() function) or the edgelist to process}

\item{intervals}{interval(s) defining the step-wise function(s). It can be specified in two different ways: (1) as a list of intervals with "widths", "K"; (2) as list specifying: a vector of integers indicating the number of steps per stepwise model ("K"), the maximum time width for intervals ("maxTime"), the minimum width for intervals ("minWidth"), the number of simulations per each K ("nsimK"). With the second specification, the function bremory::intervals() will be used.}

\item{data}{[\emph{optional}] list of exogenous variables (time varying or not)}

\item{WAIC}{[\emph{optional}] default is TRUE and is needed where WAIC weights have to be estimated}

\item{ELPD}{[\emph{optional}] default is FALSE and is TRUE if ELPD-LFO has to be calculated (PSIS approximation is used [ref. to literature here] for one-step-ahead predictions). This procedure might slow down the algorithm. [please do not run. Not yet implemented]}

\item{ncores}{[\emph{optional}] number of ncores to use for the parallelization}

\item{subset}{[\emph{optional}] first and last event index defining the sub-history on which to estimate the stepwise memory model(s)}

\item{...}{[\emph{optional}] other arguments referring to those in bremory::elpd()}
}
\value{
object of class 'smm' with attributes and methods explained in the vignette(topic = "smm", package = "bremory")
}
\description{
A function for the estimation of one or more step-wise models (at the moment working without remstats package)
}
\examples{

# examples here
}
