% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticsREH.R
\name{statisticsREH}
\alias{statisticsREH}
\title{statisticsREH}
\usage{
statisticsREH(env = globalenv(), effects = NULL, K_range = c(2:6))
}
\arguments{
\item{env}{environment where the user is currently working}

\item{effects}{list of two objects: effects$endogenous is a vector of strings with specified the effects of interest; effects$exogenous is an array [M*dyads*exogenous_effects]}

\item{K_range}{range as to the number of intervals that are going to be generated.}
}
\value{
size of the specified environment.
}
\description{
A function which creates an environment to...
}
